requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
crewai==0.28.0
openai==1.12.0
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.29.0
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.24.0
chromadb==0.4.22
openai-whisper==20231117
requests==2.31.0
httpx==0.25.0
pyyaml==6.0.1
click==8.1.7
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
isort==5.13.0
